+++ bindings/imgui_bundle/python_backends/examples/example_python_backend_pyglet.py
@@ -3,7 +3,7 @@
 
 from __future__ import absolute_import
 from imgui_bundle import imgui
-from imgui_bundle.python_backends_wip import pyglet_backend
+from imgui_bundle.python_backends import pyglet_backend
 
 from pyglet import gl  # type: ignore
 import pyglet  # type: ignore