+++ bindings/imgui_bundle/python_backends/python_backends_disabled/glumpy_backend.py
@@ -4,7 +4,7 @@ import numpy as np
 
 import ctypes
 from ctypes import *
-from imgui_bundle.python_backends_wip.opengl_backend import BaseOpenGLRenderer
+from imgui_bundle.python_backends.opengl_backend import BaseOpenGLRenderer
 
 from glumpy import gloo, gl  # type: ignore
 
@@ -14,7 +14,7 @@ from glumpy.log import log  # type: ignore
 
 import glfw
 
-from imgui_bundle.python_backends_wip import compute_fb_scale
+from imgui_bundle.python_backends import compute_fb_scale
 
 
 class GlumpyRenderer(BaseOpenGLRenderer):